Linked-list perimeter now funcions properly. Suppressed warnings.

+++ waypoints/waypoints/point.cpp
@@ -1,6 +1,6 @@
 #include "point.h"
 
-Uav::Point::Point(double lx=0, double ly=0)
+Uav::Point::Point(int lx=0, int ly=0)
 {
 	x = lx;
 	y = ly;
@@ -10,21 +10,21 @@ Uav::Point::Point()
 	x = 0;
 	y = 0;
 }
-void Uav::Point::setX(double lx)
+void Uav::Point::setX(int lx)
 {
 	x = lx;
 }
-void Uav::Point::setY(double ly)
+void Uav::Point::setY(int ly)
 {
 	y = ly;
 }
 
-double Uav::Point::getX()
+int Uav::Point::getX()
 {
 	return x;
 }
 
-double Uav::Point::getY()
+int Uav::Point::getY()
 {
 	return y;
 }
